feat: 메인화면에서 캠퍼스 선택 시 마지막 지도 중심 좌표 변경 - prevMapCenter.ts에서 lastMapCenter.ts로 이름 변경
